+++ config/config.production.js
@@ -1,7 +1,6 @@
 module.exports =
 { port: process.env.PORT,
   database: process.env.MONGO_URI,
-  bitgoUser: process.env.BITGO_USER,
   bitgoPassword: process.env.BITGO_PASSWORD,
   bitgoAccessToken: process.env.BITGO_TOKEN,
   login: