+++ config/config.test.js
@@ -1,7 +1,6 @@
 module.exports =
 { port: process.env.PORT || 3000,
   database: process.env.MONGO_URI || 'localhost/test',
-  bitgoUser: 'whoareyou@brave.com',
   bitgoPassword: 'YourPassword',
   bitgoAccessToken: 'YourAPIToken',
   login: