+++ src/controllers/helper.js
@@ -1,8 +1,9 @@
 var boom = require('boom')
-var crypto = require('crypto')
-var ecdsa = require('eccrypto')
+var createHash = require('crypto').createHash
+var Elliptic = require('elliptic').ec
+var ec = new Elliptic('p256')
 var Joi = require('joi')
-// var timestamp = require('monolithic-timestamp')
+// var timestamp = require('monotonic-timestamp')
 
 var exports = {}
 
@@ -10,80 +11,57 @@ exports.userId2stats = async function (runtime, userId) {
   var user
   var users = runtime.db.get('users')
 
-  user = await users.findOne({ userId: userId }, { statAdReplaceCount: true })
+  user = await users.findOne({ userId: userId }, { replacements: true })
 
-  return { replacements: user ? user.statAdReplaceCount : 0 }
+  return { replacements: user ? user.replacements : 0 }
 }
 
 exports.sessionId2stats = async function (runtime, userId, sessionId) {
-  var session
-  var sessions = runtime.db.get('sessions')
+  var replacements = runtime.db.get('replacements')
+  var session = await replacements.findOne({ userId: userId, sessionId: sessionId }, { replacements: true })
 
-  session = await sessions.findOne({ userId: userId, sessionId: sessionId }, { statAdReplaceCount: true })
-
-  return { replacements: session ? session.statAdReplaceCount : 0 }
-}
-
-/*
-var to_hex = function (bs) {
-  var i
-  var encoded = []
-
-  for (i = 0; i < bs.length; i++) {
-    encoded.push('0123456789abcdef'[(bs[i] >> 4) & 15])
-    encoded.push('0123456789abcdef'[bs[i] & 15])
-  }
-  return encoded.join('')
-}
- */
-
-var from_hex = function (s) {
-  var i, result
-
-  if (!s) s = ''
-  result = new Uint8Array(s.length / 2)
-  for (i = 0; i < s.length / 2; i++) result[i] = parseInt(s.substr(i * 2, 2), 16)
-
-  return result
+  return { replacements: session ? session.replacements : 0 }
 }
 
 exports.verify = async function (debug, user, data) {
-  var diff, nonce
-  var envelope = data.envelope
-  var payload = JSON.stringify(data.payload || data.intent)
+  var combo, diff, hash, nonce, r, s, signature
+  var header = data.header
+  var payload = data.payload
 
-  if (user.envelope) {
-    if (!envelope) return boom.badRequest('payload is not cryptographically-signed')
+  var u8 = function (a) { return new Uint8Array(a) }
 
-    if (user.envelope.version !== 1) {
-      return boom.badRequest('unknown user entry cryptography version: ' + JSON.stringify(user.envelope.version))
+  if (user.version) {
+    if (!header) return boom.badRequest('payload is not cryptographically-signed')
+
+    if (user.version !== 1) {
+      return boom.badRequest('unknown user entry cryptography version: ' + JSON.stringify(user.version))
     }
 
-    nonce = envelope.nonce
+    nonce = header.nonce
     if (typeof nonce === 'string') nonce = parseFloat(nonce)
-    if (isNaN(nonce)) return boom.badRequest('envelope.nonce is invalid: ' + JSON.stringify(envelope.nonce))
+    if (isNaN(nonce)) return boom.badRequest('header.nonce is invalid: ' + JSON.stringify(header.nonce))
     diff = Math.abs(new Date().getTime() - (nonce * 1000.0))
-    if (diff > 15) return boom.badData('envelope.nonce is untimely: ' + JSON.stringify(envelope.nonce))
+    // NB: 10 minutes is temporary
+    if (diff > (6000 * 1000)) return boom.badData('header.nonce is untimely: ' + JSON.stringify(header.nonce))
   } else {
-    if (envelope) return boom.badData('user entry is not cryptographically-enabled')
+    if (header) return boom.badData('user entry is not cryptographically-enabled')
 
     return null    // no user entry credentials, no data signature
   }
 
-  try {
-    await ecdsa.verify(from_hex(user.envelope.publicKey),
-                       crypto.createHash('sha256').update(user.userId + ':' + envelope.nonce + ':' + payload).digest(),
-                       from_hex(envelope.signature))
+  combo = JSON.stringify({ userId: user.userId, nonce: header.nonce, payload: payload })
+  hash = createHash('sha256').update(new Buffer(combo)).digest()
+  signature = new Buffer(header.signature, 'hex')
+  r = signature.slice(0, 32)
+  s = signature.slice(32)
 
+  try {
+    ec.verify(hash, { r: u8(r), s: u8(s) }, u8(new Buffer(user.publicKey, 'hex')))
     return null    // winner!
   } catch (ex) {
     debug('signature error', ex)
     return boom.badData('signature error', ex)
   }
-}
-
-exports.add_nonce = function (payload) {
-  return { envelope: { nonce: (new Date().getTime() / 1000.0).toString() }, payload: payload }
 }
 
 exports.add_header_schema = function (payload) {