+++ src/controllers/session.js
@@ -7,6 +7,21 @@ var underscore = require('underscore')
 
 var v1 = {}
 
+var sessionSchema = Joi.object().keys({
+  userId: Joi.string().guid().required().description('the identity of the user entry'),
+  sessionId: Joi.string().guid().required().description('the identity of the session'),
+  type: Joi.string().min(1).required().description('the name of the type'),
+  header: Joi.object({
+    signature: Joi.string().required().description('a digital signature calculated over userId:nonce:JSON.stringify(payload)'),
+    nonce: Joi.string().required().description('a time-based, monotonically-increasing value')
+  }).required(),
+  payload: Joi.object().required().keys({
+    iv: Joi.string().min(1).required().description('a once-only initialization vector'),
+    encryptedData: Joi.string().min(1).required()
+  }),
+  timestamp: Joi.string().required().description('an opaque, monotonically-increasing value')
+})
+
 /*
    GET /v1/users/{userId}/sessions[? [since={milliseconds since epoch}&] [limit={positive integer}] ]
        defaults to since=0 limit=100
@@ -17,10 +32,11 @@ v1.getSessions =
 { handler: function (runtime) {
   return async function (request, reply) {
     var entries, modifiers, query, result
-    var userId = request.params.userId
+    var userId = request.params.userId.toUpperCase()
     var limit = parseInt(request.query.limit, 10)
     var timestamp = request.query.timestamp
-    var sessionState = runtime.db.get('session_state')
+    var type = request.query.type
+    var sessions = runtime.db.get('sessions')
 
     try { timestamp = (timestamp || 0) ? bson.Timestamp.fromString(timestamp) : bson.Timestamp.ZERO } catch (ex) {
       return reply(boom.badRequest('invalid value for the timestamp parameter: ' + timestamp))
@@ -28,16 +44,17 @@ v1.getSessions =
 
     if (isNaN(limit) || (limit > 100)) limit = 100
     query = { userId: userId, timestamp: { $gte: timestamp } }
+    if (type) query.type = type
     modifiers = { sort: { timestamp: 1 } }
 
-    entries = await sessionState.find(query, underscore.extend({ limit: limit }, modifiers))
+    entries = await sessions.find(query, underscore.extend({ limit: limit }, modifiers))
     result = []
     entries.forEach(entry => {
       result.push(underscore.extend(underscore.omit(entry, '_id', 'timestamp'),
                                     { timestamp: entry.timestamp.toString() }))
     })
 
-    reply(helper.add_nonce(result))
+    reply(helper.add_nonce_data(result))
   }
 },
 
@@ -49,20 +66,13 @@ v1.getSessions =
     { params: { userId: Joi.string().guid().required().description('the identity of the user entry') },
       query:
       { timestamp: Joi.string().regex(/^[0-9]+$/).optional().description('an opaque, monotonically-increasing value'),
-        limit: Joi.number().positive().optional().description('the maximum number of entries to return')
+        limit: Joi.number().positive().optional().description('the maximum number of entries to return'),
+        type: Joi.string().min(1).optional().description('the name of the type')
       }
     },
 
   response: {
-// FIXME
-    schema: Joi.array().items(Joi.object().keys({
-      sessionId: Joi.string().guid().required().description('the identity of the session'),
-      type: Joi.string().min(1).required().description('the name of the type'),
-      userId: Joi.string().guid().required().description('the identity of the user entry'),
-      envelope: Joi.any().required(),
-      payload: Joi.object().required().description('an opaque JSON object'),
-      timestamp: Joi.string().required().description('an opaque, monotonically-increasing value')
-    }))
+    schema: helper.add_nonce_schema(Joi.array().items(sessionSchema))
 /*
     status: {
       400: Joi.object({
@@ -83,11 +93,11 @@ v1.getTypes =
 { handler: function (runtime) {
   return async function (request, reply) {
     var entries, modifiers, query, result
-    var userId = request.params.userId
-    var sessionId = request.params.sessionId
+    var userId = request.params.userId.toUpperCase()
+    var sessionId = request.params.sessionId.toUpperCase()
     var limit = parseInt(request.query.limit, 10)
     var timestamp = request.query.timestamp
-    var sessionState = runtime.db.get('session_state')
+    var sessions = runtime.db.get('sessions')
 
     try { timestamp = (timestamp || 0) ? bson.Timestamp.fromString(timestamp) : bson.Timestamp.ZERO } catch (ex) {
       return reply(boom.badRequest('invalid value for the timestamp parameter: ' + timestamp))
@@ -97,14 +107,14 @@ v1.getTypes =
     query = { userId: userId, sessionId: sessionId, timestamp: { $gte: timestamp } }
     modifiers = { sort: { timestamp: 1 } }
 
-    entries = await sessionState.find(query, underscore.extend({ limit: limit }, modifiers))
+    entries = await sessions.find(query, underscore.extend({ limit: limit }, modifiers))
     result = []
     entries.forEach(entry => {
       result.push(underscore.extend(underscore.omit(entry, '_id', 'timestamp'),
                                     { timestamp: entry.timestamp.toString() }))
     })
 
-    reply(helper.add_nonce(result))
+    reply(helper.add_nonce_data(result))
   }
 },
 
@@ -124,15 +134,7 @@ v1.getTypes =
     },
 
   response: {
-// FIXME
-    schema: Joi.array().items(Joi.object().keys({
-      sessionId: Joi.string().guid().required().description('the identity of the session'),
-      type: Joi.string().min(1).required().description('the name of the type'),
-      userId: Joi.string().guid().required().description('the identity of the user entry'),
-      envelope: Joi.any().required(),
-      payload: Joi.object().required().description('an opaque JSON object'),
-      timestamp: Joi.string().required().description('an opaque, monotonically-increasing value')
-    }))
+    schema: helper.add_nonce_schema(Joi.array().items(sessionSchema))
 /*
     status: {
       400: Joi.object({
@@ -151,16 +153,16 @@ v1.getSessionType =
 { handler: function (runtime) {
   return async function (request, reply) {
     var result
-    var userId = request.params.userId
-    var sessionId = request.params.sessionId
+    var userId = request.params.userId.toUpperCase()
+    var sessionId = request.params.sessionId.toUpperCase()
     var type = request.params.type
-    var sessionState = runtime.db.get('session_state')
+    var sessions = runtime.db.get('sessions')
 
-    result = await sessionState.findOne({ userId: userId, sessionId: sessionId, type: type })
-    if (!result) { return reply(boom.notFound('session/type entry does not exist: ' + sessionId + '/' + type)) }
+    result = await sessions.findOne({ userId: userId, sessionId: sessionId, type: type })
+    if (!result) { return reply(boom.notFound('user/session/type entry does not exist')) }
     result = underscore.extend(underscore.omit(result, '_id', 'timestamp'), { timestamp: result.timestamp.toString() })
 
-    reply(helper.add_nonce(result))
+    reply(helper.add_nonce_data(result))
   }
 },
 
@@ -176,19 +178,11 @@ v1.getSessionType =
     },
 
   response: {
-// FIXME
-    schema: Joi.object({
-      sessionId: Joi.string().guid().required().description('the identity of the session'),
-      type: Joi.string().min(1).required().description('the name of the type'),
-      userId: Joi.string().guid().required().description('the identity of the user entry'),
-      envelope: Joi.any().required(),
-      payload: Joi.object().required().description('an opaque JSON object'),
-      timestamp: Joi.string().required().description('an opaque, monotonically-increasing value')
-    })
+    schema: helper.add_nonce_schema(sessionSchema)
 /*
     status: {
       404: Joi.object({
-        boomlet: Joi.string().required().description('ad-manifest entry does not exist')
+        boomlet: Joi.string().required().description('session/type entry does not exist')
       })
     }
  */
@@ -205,11 +199,11 @@ v1.putSessionType =
   return async function (request, reply) {
     var result, state, user
     var debug = braveHapi.debug(module, request)
-    var userId = request.params.userId
-    var sessionId = request.params.sessionId
+    var userId = request.params.userId.toUpperCase()
+    var sessionId = request.params.sessionId.toUpperCase()
     var type = request.params.type
     var users = runtime.db.get('users')
-    var sessionState = runtime.db.get('session_state')
+    var sessions = runtime.db.get('sessions')
 
     user = await users.findOne({ userId: userId })
     if (!user) { return reply(boom.notFound('user entry does not exist: ' + userId)) }
@@ -220,9 +214,9 @@ v1.putSessionType =
     state = { $currentDate: { timestamp: { $type: 'timestamp' } },
               $set: request.payload
             }
-    await sessionState.update({ userId: userId, sessionId: sessionId, type }, state, { upsert: true })
+    await sessions.update({ userId: userId, sessionId: sessionId, type }, state, { upsert: true })
 
-    reply(helper.add_nonce({}))
+    reply(helper.add_nonce_data({}))
   }
 },
 
@@ -235,15 +229,14 @@ v1.putSessionType =
         sessionId: Joi.string().guid().required().description('the identity of the session'),
         type: Joi.string().min(1).required().description('the name of the type')
       },
-      payload: Joi.object().keys({
-        envelope: Joi.any().required(),
-        payload: Joi.string().required().description('the encrypted payload')
-      })
+      payload: helper.add_header_schema(Joi.object().keys({
+        iv: Joi.string().min(1).required().description('a once-only initialization vector'),
+        encryptedData: Joi.string().min(1).required()
+      }))
     },
 
   response: {
-// FIXME
-    schema: Joi.any()
+    schema: helper.add_nonce_schema(Joi.any())
 /*
     status: {
       400: Joi.object({
@@ -253,7 +246,7 @@ v1.putSessionType =
         boomlet: Joi.string().required().description('unknown user entry cryptography version')
       }),
       400: Joi.object({
-        boomlet: Joi.string().required().description('envelope.nonce is invalid')
+        boomlet: Joi.string().required().description('header.nonce is invalid')
       }),
       404: Joi.object({
         boomlet: Joi.string().required().description('user entry does not exist')
@@ -262,7 +255,7 @@ v1.putSessionType =
         boomlet: Joi.string().required().description('user entry is not cryptographically-enabled')
       }),
       422: Joi.object({
-        boomlet: Joi.string().required().description('envelope.nonce is untimely')
+        boomlet: Joi.string().required().description('header.nonce is untimely')
       }),
       422: Joi.object({
         boomlet: Joi.string().required().description('signature error')
@@ -275,20 +268,72 @@ v1.putSessionType =
   }
 }
 
+/*
+   DELETE /v1/users/{userId}/sessions/{sessionId}/types/{type}
+ */
+
+v1.delete =
+{ handler: function (runtime) {
+  return async function (request, reply) {
+    var result, user
+    var debug = braveHapi.debug(module, request)
+    var userId = request.params.userId.toUpperCase()
+    var sessionId = request.params.sessionId.toUpperCase()
+    var type = request.params.type
+    var users = runtime.db.get('users')
+    var sessions = runtime.db.get('sessions')
+
+    user = await users.findOne({ userId: userId })
+    if (!user) { return reply(boom.notFound('user entry does not exist: ' + userId)) }
+
+    result = await helper.verify(debug, user, request.payload)
+    if (result) return reply(result)
+
+    result = await sessions.remove({ userId: userId, sessionId: sessionId, type: type })
+
+    reply().code(204)
+  }
+},
+
+  description: 'Delete session information for a particular user entry',
+  notes: 'For the purpose authentication the HTTP body must be present and contain a header/payload pairing, the payload may be any JSON value.',
+  tags: ['api'],
+
+  validate:
+    { params:
+      { userId: Joi.string().guid().required().description('the identity of the user entry'),
+        sessionId: Joi.string().guid().required().description('the identity of the session'),
+        type: Joi.string().min(1).required().description('the name of the type')
+      },
+      payload: helper.add_header_schema(Joi.any())
+    },
+
+  response: {
+/*
+    status: {
+      404: Joi.object({
+        boomlet: Joi.string().required().description('user entry does not exist')
+      })
+    }
+ */
+  }
+}
+
 module.exports.routes =
 [ braveHapi.routes.async().path('/v1/users/{userId}/sessions').config(v1.getSessions),
   braveHapi.routes.async().path('/v1/users/{userId}/sessions/{sessionId}/types').config(v1.getTypes),
   braveHapi.routes.async().path('/v1/users/{userId}/sessions/{sessionId}/types/{type}').config(v1.getSessionType),
-  braveHapi.routes.async().put().path('/v1/users/{userId}/sessions/{sessionId}/types/{type}').config(v1.putSessionType)
+  braveHapi.routes.async().put().path('/v1/users/{userId}/sessions/{sessionId}/types/{type}').config(v1.putSessionType),
+  braveHapi.routes.async().delete().path('/v1/users/{userId}/sessions/{sessionId}/types/{type}').config(v1.delete)
 ]
 
 module.exports.initialize = async function (debug, runtime) {
   runtime.db.checkIndices(debug,
-  [ { category: runtime.db.get('session_state'),
-      name: 'session_state',
+  [ { category: runtime.db.get('sessions'),
+      name: 'sessions',
       property: 'sessionId',
       empty: { userId: '', sessionId: '', type: '', timestamp: bson.Timestamp.ZERO },
-      others: [ { userId: 0 }, { sessionId: 0 }, { type: 1 }, { timestamp: 1 } ]
+      others: [ { userId: 1 }, { sessionId: 1 }, { type: 1 }, { timestamp: 1 } ]
     }
   ])
 }